refactor: controller 패키지 생성
ScriptController와 SwaggerScriptInterface를 controller 패키지로 이동
