+++ src/main/java/com/twentythree/peech/script/controller/ScriptController.java
@@ -1,4 +1,4 @@
-package com.twentythree.peech.script;
+package com.twentythree.peech.script.controller;
 
 import com.twentythree.peech.script.dto.ParagraphResponseDTO;
